Split apart the Scripture and ScriptureList components into their own files.
