+++ packages/apolloschurchapp/src/ui/Scripture/ScriptureList.js
@@ -1,0 +1,11 @@
+const ScriptureList = ({ scripture, commas = true }) => {
+  if (!scripture && commas) return '';
+  if (!scripture) return [];
+
+  const combo = scripture.map(({ reference }) => `${reference}`);
+
+  if (commas) return combo.join(', ');
+  return combo;
+};
+
+export default ScriptureList;
